+++ packages/node-core/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-- Fixed issue Buffer within sandbox could not concat Buffer and Uint8Array
+- Fixed issue Buffer within sandbox could not concat Buffer and Uint8Array, also injected `atob`
 
 ### Added
 - Expose option to add workerData to workers (#2372)